refactor: 에러 처리 컴포넌트 공통화​​
