--- src/components/MovieListBox/ErrorBox/MovieListErrorBox.ts
@@ -1,32 +1,0 @@
-import "./movieListErrorBox.css";
-
-export const addMovieListErrorBox = (text: string) => {
-  const $movieListSection = document.querySelector(
-    ".movie-list-section"
-  ) as HTMLElement;
-  $movieListSection.replaceChildren($MovieListErrorBox({ text }));
-};
-
-interface ErrorBoxProps {
-  text: string;
-}
-
-const $MovieListErrorBox = ({ text }: ErrorBoxProps) => {
-  const $errorPlanet = createElement("img", {
-    src: "./empty-planet.svg",
-    className: "empty-planet",
-    alt: text,
-  });
-  const $emptyText = createElement("h2", {
-    textContent: text,
-  });
-
-  const $box = createElement("div", {
-    className: "error-box",
-  });
-  $box.append($errorPlanet, $emptyText);
-
-  return $box;
-};
-
-export default $MovieListErrorBox;
